requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
openai==1.3.5
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
plotly==5.18.0
kaleido==0.2.1
pydantic==2.5.1
pydantic-settings==2.0.3
